get table columns from a hook, other small tweaks
