+++ src/components/Footer.js
@@ -58,7 +58,7 @@ const Footer = () => {
       }
       secondary={
         <Logo
-          heading={<h3 className="usa-footer__logo-heading">Name of Agency</h3>}
+          heading={<h3 className="usa-footer__logo-heading">OpenNC</h3>}
           image={
             <img
               alt="img alt text"
