+++ src/hooks/useTableColumns.js
@@ -1,0 +1,53 @@
+import { useMemo } from 'react'
+
+export const useTableColumns = () => {
+  const donorColumns = useMemo(
+    () => [
+      {
+        Header: 'Donor Name',
+        accessor: 'name',
+      },
+      {
+        Header: 'City/State',
+        accessor: ({ city, state }) => city + ', ' + state,
+      },
+      {
+        Header: 'Type',
+        accessor: 'type',
+      },
+      {
+        Header: 'Profession',
+        accessor: 'profession',
+      },
+      {
+        Header: 'Total Donations',
+        accessor: 'total',
+      },
+    ],
+    []
+  )
+
+  const candidateColumns = useMemo(
+    () => [
+      {
+        Header: 'Name',
+        accessor: 'candidate_full_name',
+      },
+      {
+        Header: 'Party',
+        accessor: 'party',
+      },
+      {
+        Header: 'Contest',
+        accessor: ({ office, juris }) =>
+          juris ? `${office} ${juris}` : office,
+      },
+    ],
+    []
+  )
+
+  return {
+    donorColumns,
+    candidateColumns,
+  }
+}
